Client - minor fixes
* Policy max value can now be a double rather than an int
* Store label does not render in the store products view page
* Add policy step of the min and max values controles now changes based on the rule type
* Now displaying policy min and max values with or without decimal point based on rule type

+++ TradingSystem/TradingSystem.WebApi/DTO/Store/Policy/PolicyParamsDTO.cs
@@ -20,9 +20,8 @@ namespace TradingSystem.WebApi.DTO.Store.Policy
         public string? Category { get; set; }
         public Guid? ProductId { get; set; }
 
-
-        public int? MinValue { get; set; }
-        public int? MaxValue { get; set; }
+        public double? MinValue { get; set; }
+        public double? MaxValue { get; set; }
 
         public DateTime? StartDate { get; set; }
         public DateTime? EndDate { get; set; }
